clean up and simplify
