+++ src/components/pages/portfolio/types/api.ts
@@ -152,13 +152,6 @@ export const portfolioActivityResponseSchema = z.object({
   version: z.enum(['all', 'v2', 'v3']).optional().default('all'),
   limit: z.number(),
   offset: z.number(),
-  facets: z
-    .object({
-      chainIds: z.array(z.number())
-    })
-    .nullable()
-    .optional()
-    .default(null),
   pageInfo: z.object({
     hasMore: z.boolean(),
     nextOffset: z.number().nullable()
@@ -171,10 +164,6 @@ export const portfolioActivityFacetsResponseSchema = z.object({
   version: z.enum(['all', 'v2', 'v3']).optional().default('all'),
   facets: z.object({
     chainIds: z.array(z.number())
-  }),
-  pageInfo: z.object({
-    hasMore: z.boolean(),
-    nextOffsetPerSource: z.number().nullable()
   })
 })
